Projects/App_009_001/requirements.txt
lxml==4.9.3
matplotlib==3.5.1
numpy==1.24.4
pandas==2.0.3
pickleshare==0.7.5
pillow==9.0.1
scikit-learn==1.0.2
scipy==1.11.1
shap==0.42.1
streamlit==1.26.0